Move LoggedEvent model to amplitude package.
