+++ amplitude/management/commands/export_to_amplitude.py
@@ -11,7 +11,7 @@ from onboarding.models import OnboardingInfo
 from amplitude.models import Sync, SYNC_CHOICES
 from amplitude.api import AmpEvent, AmpEventUploader, EPOCH
 from evictionfree.models import SubmittedHardshipDeclaration
-from frontend.models import LoggedEvent
+from amplitude.models import LoggedEvent
 
 
 class Synchronizer(ABC):
@@ -93,7 +93,7 @@ class EfnySynchronizer(Synchronizer):
             )
 
 
-class FrontendSynchronizer(Synchronizer):
+class AmplitudeLoggedEventSynchronizer(Synchronizer):
     @classmethod
     def iter_events(cls, last_synced_at: datetime.datetime) -> Iterator[AmpEvent]:
         qs = LoggedEvent.objects.filter(created_at__gte=last_synced_at).select_related("user")
@@ -108,7 +108,7 @@ class FrontendSynchronizer(Synchronizer):
 SYNCHRONIZERS: Dict[str, Synchronizer] = {
     SYNC_CHOICES.USERS_V2: UserSynchronizer(),
     SYNC_CHOICES.EVICTIONFREE: EfnySynchronizer(),
-    SYNC_CHOICES.FRONTEND: FrontendSynchronizer(),
+    SYNC_CHOICES.AMPLITUDE: AmplitudeLoggedEventSynchronizer(),
 }
