+++ common-data/amplitude-sync-choices.json
@@ -2,5 +2,5 @@
   ["USERS", "Users v1 (deprecated)"],
   ["USERS_V2", "Users v2"],
   ["EVICTIONFREE", "EvictionFreeNY.org submitted hardship declarations"],
-  ["FRONTEND", "Logged front-end events"]
+  ["AMPLITUDE", "Logged events for Amplitude"]
 ]
